[기록] Realm으로 앱이 죽는 이슈 해결 -> SDK의 Realm과 App단의 Realm이 겹쳐서 발생한 문제
